FIX: build error - cast from pointer to integer of different size

+++ json.c
@@ -42,6 +42,7 @@ const struct _json_value json_value_none;
 #include <string.h>
 #include <ctype.h>
 #include <math.h>
+#include <stdint.h>
 
 typedef unsigned int json_uchar;
 
@@ -153,7 +154,7 @@ static int new_value (json_state * state,
             values_size = sizeof (*value->u.object.values) * value->u.object.length;
 
             if (! (value->u.object.values = (json_object_entry *) json_alloc
-                  (state, values_size + ((unsigned long) value->u.object.values), 0)) )
+                  (state, values_size + ((uintptr_t) value->u.object.values), 0)) )
             {
                return 0;
             }
